third_party/WebKit/LayoutTests/inspector/tracing/timeline-load-event.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/timeline-test.js"></script>
<script>

function test()
{
    InspectorTest.startTimeline(function() { InspectorTest.reloadPage(pageReloaded); });

    function pageReloaded()
    {
        InspectorTest.stopTimeline(finish);
    }

    function finish()
    {
        InspectorTest.addResult("Model records:");
        InspectorTest.printTimelineRecords("MarkDOMContent");
        InspectorTest.printTimelineRecords("MarkLoad");
        InspectorTest.addResult("Timestamp records:");
        InspectorTest.printTimestampRecords("MarkDOMContent");
        InspectorTest.printTimestampRecords("MarkLoad");
        InspectorTest.completeTest();
    }
}

</script>
</head>

<body onload="runTest()">
<p>
Tests the load event.
</p>

</body>
</html>
